+++ packages/koenig-lexical/src/utils/getTopLevelNativeElement.js
@@ -1,0 +1,8 @@
+export function getTopLevelNativeElement(node) {
+    if (node.nodeType === Node.TEXT_NODE) {
+        node = node.parentNode;
+    }
+
+    const selector = '[data-lexical-editor] > *';
+    return node.closest(selector);
+}
