+++ packages/koenig-lexical/src/utils/index.js
@@ -1,3 +1,5 @@
 // publicly exported util functions
 export * from './$isAtStartOfDocument';
 export * from './$selectDecoratorNode';
+export * from './$isAtTopOfNode';
+export * from './getTopLevelNativeElement';
